Move Icon and EmptyIcon class to Drupal\at_base\Icon namespace.
