+++ lib/Icon/IconInterface.php
@@ -7,7 +7,7 @@ interface IconInterface {
    * Get Icon instance with information to generate icon tag.
    *
    * @param string $css_code
-   * @return \Drupal\at_base\Icon
+   * @return \Drupal\at_base\Icon\Icon
    */
   public function get($css_code);
 }
